+++ glean/src/index/qt.ts
@@ -17,6 +17,7 @@ import EventMetricType from "../core/metrics/types/event.js";
 import LabeledMetricType from "../core/metrics/types/labeled.js";
 import QuantityMetricType from "../core/metrics/types/quantity.js";
 import StringMetricType from "../core/metrics/types/string.js";
+import StringListMetricType from "../core/metrics/types/string_list";
 import TimespanMetricType from "../core/metrics/types/timespan.js";
 import UUIDMetricType from "../core/metrics/types/uuid.js";
 import URLMetricType from "../core/metrics/types/url.js";
@@ -123,6 +124,7 @@ export default {
     LabeledMetricType,
     QuantityMetricType,
     StringMetricType,
+    StringListMetricType,
     TimespanMetricType,
     UUIDMetricType,
     URLMetricType